+++ packages/orchestrator_core/lib/orchestrator_core.dart
@@ -57,3 +57,4 @@ export 'src/utils/job_result.dart';
 export 'src/utils/event_extensions.dart';
 export 'src/utils/state_patterns.dart';
 export 'src/utils/orchestrator_helpers.dart';
+export 'src/utils/saga_flow.dart';